requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
dj-database-url==2.0.0
Django==4.0.4
django-cors-headers==4.2.0
django-environ==0.11.2

django-filter==23.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
gunicorn==21.2.0
importlib-resources==6.0.1
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
packaging==23.1
Pillow==10.0.1
pkgutil_resolve_name==1.3.10
psycopg2==2.9.7
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.3
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
zipp==3.16.2
load-dotenv==0.1.0